Convenience function get_const_tensor (#45)
Get the constant tensor from a value, if it exists.

A constant tensor can be obtained if the value has a ``const_value`` set
(as in the case of an initializer) or if the value is produced by a
Constant node.

This function will not alter the ``const_value`` of the value, but
it will propagate the shape and type of the constant tensor to the value
if `propagate_shape_type` is set to True.

---------

Signed-off-by: Justin Chu <justinchuby@users.noreply.github.com>
Co-authored-by: codecov-ai[bot] <156709835+codecov-ai[bot]@users.noreply.github.com>